auto 0 53 409

+++ itoa.h
@@ -3,8 +3,8 @@
 
 #define BASE_Q -2
 
-size_t int_to_base(char s[], size_t space, unsigned int n, int base);
-size_t itoa(char s[], size_t space, unsigned int n);
+size_t int_to_base(char s[], size_t space, unsigned long n, int base);
+size_t itoa(char s[], size_t space, unsigned long n);
 
 size_t double_to_base(char s[], size_t space, double n, int base);
 size_t dtoa(char s[], size_t space, double n);
